wrapper: extract js evaluating fns into own module
